requirements.txt
fastapi==0.61.1
pydantic==1.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
uvicorn==0.11.8
uvloop==0.14.0
pytz==2020.1
starlette==0.13.6
#boto3==1.14.56
#botocore==1.17.56
#mangum==0.9.2
pynamodb==4.3.3